refactor: 클래스 이동, 네이밍 개선
