+++ src/main/java/bowling/domain/InProgress.java
@@ -1,0 +1,14 @@
+package bowling.domain;
+
+abstract class InProgress implements PinMarkerState {
+
+    @Override
+    public boolean isStarted() {
+        return true;
+    }
+
+    @Override
+    public boolean isCompleted() {
+        return false;
+    }
+}
